fix (BRI-21) | Вынес ачивки в отдельный компонент
