+++ packages/client/src/components/core/Achievements/achievements.module.scss
@@ -1,0 +1,15 @@
+.achievements {
+    width: 6vw;
+    margin-left: 6%;
+    margin-top: 2%;
+    background-color: rgba(255, 255, 255, 0.9);
+    border-radius: 1vw;
+}
+.achievements__icon {
+    width: 6vw;
+    height: 4vw;
+}
+.achievements__p {
+    font-size: .8vw;
+    text-align: center;
+}
