+++ packages/client/src/components/core/Achievements/achievements.tsx
@@ -1,0 +1,14 @@
+import styles from "./achievements.module.scss"
+
+type achieve = {
+    imp_path: string;
+    achieve_text:string;
+}
+
+
+export const Achievements = (props:achieve) => {
+    return <div className={styles.achievements}>
+                <img src={props.imp_path} alt="" className={styles.achievements__icon} />
+                <p className={styles.achievements_p}>{props.achieve_text}</p>
+            </div>
+}
