+++ packages/client/src/components/core/Achievements/index.tsx
@@ -1,0 +1,1 @@
+export {Achievements} from "./achievements"
